+++ sniffer/app.py
@@ -52,7 +52,7 @@ class JStationSnifferApp:
         self.js_sniffer = JStationSniffer(sys.argv[0])
         self.midi_dlg = MidiSelectDlg(self.gtk_builder, self,
                                       self.js_sniffer,
-                                      self.start_sniffing, self.quit)
+                                      self.on_connected, self.quit)
         self.midi_dlg.gtk_dlg.connect('destroy', self.quit)
 
         self.connect()
@@ -63,24 +63,26 @@ class JStationSnifferApp:
         if self.config.has_section('MIDI'):
             midi_port_in = self.config.get('MIDI', 'port_in')
             midi_port_out = self.config.get('MIDI', 'port_out')
+            sysex_channel = self.config.getint('MIDI', 'sysex_channel')
 
         if midi_port_in and midi_port_out:
-            self.js_sniffer.connect(midi_port_in, midi_port_out, 1)
+            self.js_sniffer.connect(midi_port_in, midi_port_out, sysex_channel)
             if self.js_sniffer.is_connected:
                 self.midi_dlg.set_defaults(midi_port_in, midi_port_out)
                 self.midi_dlg.set_connected()
-                self.start_sniffing(midi_port_in, midi_port_out)
+                self.on_connected(midi_port_in, midi_port_out, sysex_channel)
 
         self.midi_dlg.present()
 
-    def start_sniffing(self, midi_port_in, midi_port_out):
+    def on_connected(self, midi_port_in, midi_port_out, sysex_channel):
         if self.js_sniffer.is_connected:
             if not self.config.has_section('MIDI'):
                 self.config.add_section('MIDI')
             self.config.set('MIDI', 'port_in', midi_port_in)
             self.config.set('MIDI', 'port_out', midi_port_out)
+            self.config.set('MIDI', 'sysex_channel', '%d'%(sysex_channel))
 
-        self.js_sniffer.start_sniffer(midi_port_in, midi_port_out)
+        self.js_sniffer.start_sniffer()
         self.midi_dlg.msg_lbl.set_text('Sniffing events...')
 
     def quit(self, window=None):